moved encoders java.awt under gui
