+++ src/main/java/net/pms/encoders/MEncoderVideo.java
@@ -19,13 +19,7 @@ package net.pms.encoders;
 
 import bsh.EvalError;
 import bsh.Interpreter;
-import com.jgoodies.forms.builder.PanelBuilder;
-import com.jgoodies.forms.factories.Borders;
-import com.jgoodies.forms.layout.CellConstraints;
-import com.jgoodies.forms.layout.FormLayout;
 import com.sun.jna.Platform;
-import java.awt.*;
-import java.awt.event.*;
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -35,7 +29,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import javax.swing.*;
 import net.pms.Messages;
 import net.pms.PMS;
 import net.pms.configuration.DeviceConfiguration;
@@ -50,14 +43,10 @@ import net.pms.formats.Format;
 import net.pms.formats.v2.SubtitleType;
 import net.pms.io.*;
 import net.pms.network.HTTPResource;
-import net.pms.newgui.GuiUtil;
-import net.pms.newgui.components.CustomJButton;
 import net.pms.platform.windows.NTStatus;
 import net.pms.util.*;
 import static net.pms.util.AudioUtils.getLPCMChannelMappingForMencoder;
 import static net.pms.util.StringUtil.quoteArg;
-import org.apache.commons.configuration.event.ConfigurationEvent;
-import org.apache.commons.configuration.event.ConfigurationListener;
 import org.apache.commons.lang3.ArrayUtils;
 import static org.apache.commons.lang.BooleanUtils.isTrue;
 import static org.apache.commons.lang3.StringUtils.*;
@@ -75,31 +64,8 @@ public class MEncoderVideo extends Player {
 	public static final String KEY_MENCODER_EXECUTABLE_TYPE = "mencoder_executable_type";
 	public static final String NAME = "MEncoder Video";
 
-	private static final String COL_SPEC = "left:pref, 3dlu, p:grow, 3dlu, right:p:grow, 3dlu, p:grow, 3dlu, right:p:grow,3dlu, p:grow, 3dlu, right:p:grow,3dlu, pref:grow";
-	private static final String ROW_SPEC = "p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 9dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p";
 	private static final String REMOVE_OPTION = "---REMOVE-ME---"; // use an out-of-band option that can't be confused with a real option
 
-	private JTextField mencoderNoassScale;
-	private JTextField mencoderNoassSubpos;
-	private JTextField mencoderNoassBlur;
-	private JTextField mencoderNoassOutline;
-	private JTextField mencoderCustomOptions;
-	private JTextField subq;
-	private JCheckBox forcefps;
-	private JCheckBox yadif;
-	private JCheckBox scaler;
-	private JTextField scaleX;
-	private JTextField scaleY;
-	private JCheckBox fc;
-	private JCheckBox ass;
-	private JCheckBox skipLoopFilter;
-	private JCheckBox mencodermt;
-	private JCheckBox videoremux;
-	private JCheckBox normalizeaudio;
-	private JCheckBox noskip;
-	private JCheckBox intelligentsync;
-	private JTextField ocw;
-	private JTextField och;
 
 	private static final String[] INVALID_CUSTOM_OPTIONS = {
 		"-of",
@@ -150,370 +116,6 @@ public class MEncoderVideo extends Player {
 
 	// Not to be instantiated by anything but PlayerFactory
 	MEncoderVideo() {
-	}
-
-	public JCheckBox getNoskip() {
-		return noskip;
-	}
-
-	@Override
-	public JComponent config() {
-		// Apply the orientation for the locale
-		ComponentOrientation orientation = ComponentOrientation.getOrientation(PMS.getLocale());
-		String colSpec = FormLayoutUtil.getColSpec(COL_SPEC, orientation);
-
-		FormLayout layout = new FormLayout(colSpec, ROW_SPEC);
-		PanelBuilder builder = new PanelBuilder(layout);
-		builder.border(Borders.EMPTY);
-		builder.opaque(false);
-
-		CellConstraints cc = new CellConstraints();
-
-		JComponent cmp = builder.addSeparator(Messages.getString("GeneralSettings_SentenceCase"), FormLayoutUtil.flip(cc.xyw(1, 1, 15), colSpec, orientation));
-		cmp = (JComponent) cmp.getComponent(0);
-		cmp.setFont(cmp.getFont().deriveFont(Font.BOLD));
-
-		mencodermt = new JCheckBox(Messages.getString("EnableMultithreading"), configuration.getMencoderMT());
-		mencodermt.setContentAreaFilled(false);
-		mencodermt.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				configuration.setMencoderMT(mencodermt.isSelected());
-			}
-		});
-		mencodermt.setEnabled(Platform.isWindows() || Platform.isMac());
-		builder.add(GuiUtil.getPreferredSizeComponent(mencodermt), FormLayoutUtil.flip(cc.xy(1, 3), colSpec, orientation));
-
-		skipLoopFilter = new JCheckBox(Messages.getString("SkipLoopFilterDeblocking"), configuration.getSkipLoopFilterEnabled());
-		skipLoopFilter.setContentAreaFilled(false);
-		skipLoopFilter.setToolTipText(Messages.getString("CanDegradeQuality"));
-		skipLoopFilter.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setSkipLoopFilterEnabled((e.getStateChange() == ItemEvent.SELECTED));
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(skipLoopFilter), FormLayoutUtil.flip(cc.xyw(3, 3, 12), colSpec, orientation));
-
-		noskip = new JCheckBox(Messages.getString("AvSyncAlternativeMethod"), configuration.isMencoderNoOutOfSync());
-		noskip.setContentAreaFilled(false);
-		noskip.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderNoOutOfSync((e.getStateChange() == ItemEvent.SELECTED));
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(noskip), FormLayoutUtil.flip(cc.xy(1, 5), colSpec, orientation));
-
-		CustomJButton button = new CustomJButton(Messages.getString("CodecSpecificParametersAdvanced"));
-		button.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				JPanel codecPanel = new JPanel(new BorderLayout());
-				final JTextArea textArea = new JTextArea();
-				textArea.setText(configuration.getMencoderCodecSpecificConfig());
-				textArea.setFont(new Font("Courier", Font.PLAIN, 12));
-				JScrollPane scrollPane = new JScrollPane(textArea);
-				scrollPane.setPreferredSize(new Dimension(900, 100));
-
-				final JTextArea textAreaDefault = new JTextArea();
-				textAreaDefault.setText(DEFAULT_CODEC_CONF_SCRIPT);
-				textAreaDefault.setBackground(Color.WHITE);
-				textAreaDefault.setFont(new Font("Courier", Font.PLAIN, 12));
-				textAreaDefault.setEditable(false);
-				textAreaDefault.setEnabled(configuration.isMencoderIntelligentSync());
-				JScrollPane scrollPaneDefault = new JScrollPane(textAreaDefault);
-				scrollPaneDefault.setPreferredSize(new Dimension(900, 450));
-
-				JPanel customPanel = new JPanel(new BorderLayout());
-				intelligentsync = new JCheckBox(Messages.getString("UseApplicationDefaults"), configuration.isMencoderIntelligentSync());
-				intelligentsync.setContentAreaFilled(false);
-				intelligentsync.addItemListener(new ItemListener() {
-					@Override
-					public void itemStateChanged(ItemEvent e) {
-						configuration.setMencoderIntelligentSync((e.getStateChange() == ItemEvent.SELECTED));
-						textAreaDefault.setEnabled(configuration.isMencoderIntelligentSync());
-
-					}
-				});
-
-				JLabel label = new JLabel(Messages.getString("CustomParameters"));
-				customPanel.add(label, BorderLayout.NORTH);
-				customPanel.add(scrollPane, BorderLayout.SOUTH);
-
-				codecPanel.add(intelligentsync, BorderLayout.NORTH);
-				codecPanel.add(scrollPaneDefault, BorderLayout.CENTER);
-				codecPanel.add(customPanel, BorderLayout.SOUTH);
-
-				while (JOptionPane.showOptionDialog(SwingUtilities.getWindowAncestor((Component) PMS.get().getFrame()),
-					codecPanel, Messages.getString("EditCodecSpecificParameters"), JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE, null, null, null) == JOptionPane.OK_OPTION) {
-					String newCodecparam = textArea.getText();
-					DLNAMediaInfo fakemedia = new DLNAMediaInfo();
-					DLNAMediaAudio audio = new DLNAMediaAudio();
-					audio.setCodecA("ac3");
-					fakemedia.setCodecV("mpeg4");
-					fakemedia.setContainer("matroska");
-					fakemedia.setDuration(45d * 60);
-					audio.getAudioProperties().setNumberOfChannels(2);
-					fakemedia.setWidth(1280);
-					fakemedia.setHeight(720);
-					audio.setSampleFrequency("48000");
-					fakemedia.setFrameRate("23.976");
-					fakemedia.getAudioTracksList().add(audio);
-					String[] result = getSpecificCodecOptions(newCodecparam, fakemedia, new OutputParams(configuration), "dummy.mpg", "dummy.srt", false, true);
-
-					if (result.length > 0 && result[0].startsWith("@@")) {
-						String errorMessage = result[0].substring(2);
-						JOptionPane.showMessageDialog(
-							SwingUtilities.getWindowAncestor((Component) PMS.get().getFrame()),
-							errorMessage,
-							Messages.getString("Error"),
-							JOptionPane.ERROR_MESSAGE
-						);
-					} else {
-						configuration.setMencoderCodecSpecificConfig(newCodecparam);
-						break;
-					}
-				}
-			}
-		});
-		builder.add(button, FormLayoutUtil.flip(cc.xy(1, 11), colSpec, orientation));
-
-		forcefps = new JCheckBox(Messages.getString("ForceFramerateParsedFfmpeg"), configuration.isMencoderForceFps());
-		forcefps.setContentAreaFilled(false);
-		forcefps.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderForceFps(e.getStateChange() == ItemEvent.SELECTED);
-			}
-		});
-
-		builder.add(GuiUtil.getPreferredSizeComponent(forcefps), FormLayoutUtil.flip(cc.xyw(1, 7, 2), colSpec, orientation));
-
-		yadif = new JCheckBox(Messages.getString("DeinterlaceFilter_Sentencecase"), configuration.isMencoderYadif());
-		yadif.setContentAreaFilled(false);
-		yadif.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderYadif(e.getStateChange() == ItemEvent.SELECTED);
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(yadif), FormLayoutUtil.flip(cc.xyw(3, 7, 7), colSpec, orientation));
-
-		scaler = new JCheckBox(Messages.getString("ChangeVideoResolution"));
-		scaler.setContentAreaFilled(false);
-		scaler.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderScaler(e.getStateChange() == ItemEvent.SELECTED);
-				scaleX.setEnabled(configuration.isMencoderScaler());
-				scaleY.setEnabled(configuration.isMencoderScaler());
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(scaler), FormLayoutUtil.flip(cc.xyw(3, 5, 6), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("Width"), FormLayoutUtil.flip(cc.xy(9, 5, CellConstraints.RIGHT, CellConstraints.CENTER), colSpec, orientation));
-		scaleX = new JTextField("" + configuration.getMencoderScaleX());
-		scaleX.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				try {
-					configuration.setMencoderScaleX(Integer.parseInt(scaleX.getText()));
-				} catch (NumberFormatException nfe) {
-					LOGGER.debug("Could not parse scaleX from \"" + scaleX.getText() + "\"");
-				}
-			}
-		});
-		builder.add(scaleX, FormLayoutUtil.flip(cc.xy(11, 5), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("Height"), FormLayoutUtil.flip(cc.xy(13, 5, CellConstraints.RIGHT, CellConstraints.CENTER), colSpec, orientation));
-		scaleY = new JTextField("" + configuration.getMencoderScaleY());
-		scaleY.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				try {
-					configuration.setMencoderScaleY(Integer.parseInt(scaleY.getText()));
-				} catch (NumberFormatException nfe) {
-					LOGGER.debug("Could not parse scaleY from \"" + scaleY.getText() + "\"");
-				}
-			}
-		});
-		builder.add(scaleY, FormLayoutUtil.flip(cc.xy(15, 5), colSpec, orientation));
-
-		if (configuration.isMencoderScaler()) {
-			scaler.setSelected(true);
-		} else {
-			scaleX.setEnabled(false);
-			scaleY.setEnabled(false);
-		}
-
-		videoremux = new JCheckBox(Messages.getString("RemuxVideosTsmuxer"), configuration.isMencoderMuxWhenCompatible());
-		videoremux.setContentAreaFilled(false);
-		videoremux.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderMuxWhenCompatible((e.getStateChange() == ItemEvent.SELECTED));
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(videoremux), FormLayoutUtil.flip(cc.xyw(1, 9, 13), colSpec, orientation));
-
-		normalizeaudio = new JCheckBox(Messages.getString("NormalizeAudioVolume"), configuration.isMEncoderNormalizeVolume());
-		normalizeaudio.setContentAreaFilled(false);
-		normalizeaudio.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMEncoderNormalizeVolume((e.getStateChange() == ItemEvent.SELECTED));
-			}
-		});
-		// Uncomment this if volume normalizing in MEncoder is ever fixed.
-		// builder.add(normalizeaudio, FormLayoutUtil.flip(cc.xyw(1, 13, 13), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("CustomOptionsVf"), FormLayoutUtil.flip(cc.xy(1, 15), colSpec, orientation));
-		mencoderCustomOptions = new JTextField(configuration.getMencoderCustomOptions());
-		mencoderCustomOptions.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderCustomOptions(mencoderCustomOptions.getText());
-			}
-		});
-		builder.add(mencoderCustomOptions, FormLayoutUtil.flip(cc.xyw(3, 15, 13), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("AddBordersOverscanCompensation"), FormLayoutUtil.flip(cc.xy(1, 17), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("Width") + " (%)", FormLayoutUtil.flip(cc.xy(1, 17, CellConstraints.RIGHT, CellConstraints.CENTER), colSpec, orientation));
-		ocw = new JTextField(configuration.getMencoderOverscanCompensationWidth());
-		ocw.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderOverscanCompensationWidth(ocw.getText());
-			}
-		});
-		builder.add(ocw, FormLayoutUtil.flip(cc.xy(3, 17), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("Height") + " (%)", FormLayoutUtil.flip(cc.xy(5, 17), colSpec, orientation));
-		och = new JTextField(configuration.getMencoderOverscanCompensationHeight());
-		och.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderOverscanCompensationHeight(och.getText());
-			}
-		});
-		builder.add(och, FormLayoutUtil.flip(cc.xy(7, 17), colSpec, orientation));
-
-		cmp = builder.addSeparator(Messages.getString("SubtitlesSettings"), FormLayoutUtil.flip(cc.xyw(1, 19, 15), colSpec, orientation));
-		cmp = (JComponent) cmp.getComponent(0);
-		cmp.setFont(cmp.getFont().deriveFont(Font.BOLD));
-
-		builder.addLabel(Messages.getString("PlaintextSubtitlesSettings"), FormLayoutUtil.flip(cc.xy(1, 27), colSpec, orientation));
-		builder.addLabel(Messages.getString("FontScale"), FormLayoutUtil.flip(cc.xy(1, 27, CellConstraints.RIGHT, CellConstraints.CENTER), colSpec, orientation));
-
-		mencoderNoassScale = new JTextField(configuration.getMencoderNoAssScale());
-		mencoderNoassScale.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderNoAssScale(mencoderNoassScale.getText());
-			}
-		});
-
-		builder.addLabel(Messages.getString("FontOutline"), FormLayoutUtil.flip(cc.xy(5, 27), colSpec, orientation));
-
-		mencoderNoassOutline = new JTextField(configuration.getMencoderNoAssOutline());
-		mencoderNoassOutline.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderNoAssOutline(mencoderNoassOutline.getText());
-			}
-		});
-
-		builder.addLabel(Messages.getString("FontBlur"), FormLayoutUtil.flip(cc.xy(9, 27), colSpec, orientation));
-
-		mencoderNoassBlur = new JTextField(configuration.getMencoderNoAssBlur());
-		mencoderNoassBlur.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderNoAssBlur(mencoderNoassBlur.getText());
-			}
-		});
-
-		builder.addLabel(Messages.getString("MarginPercentage"), FormLayoutUtil.flip(cc.xy(13, 27), colSpec, orientation));
-
-		mencoderNoassSubpos = new JTextField(configuration.getMencoderNoAssSubPos());
-		mencoderNoassSubpos.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderNoAssSubPos(mencoderNoassSubpos.getText());
-			}
-		});
-
-		builder.add(mencoderNoassScale, FormLayoutUtil.flip(cc.xy(3, 27), colSpec, orientation));
-		builder.add(mencoderNoassOutline, FormLayoutUtil.flip(cc.xy(7, 27), colSpec, orientation));
-		builder.add(mencoderNoassBlur, FormLayoutUtil.flip(cc.xy(11, 27), colSpec, orientation));
-		builder.add(mencoderNoassSubpos, FormLayoutUtil.flip(cc.xy(15, 27), colSpec, orientation));
-
-		ass = new JCheckBox(Messages.getString("UseAssSubtitlesStyling"), configuration.isMencoderAss());
-		ass.setContentAreaFilled(false);
-		ass.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				if (e != null) {
-					configuration.setMencoderAss(e.getStateChange() == ItemEvent.SELECTED);
-				}
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(ass), FormLayoutUtil.flip(cc.xy(1, 23), colSpec, orientation));
-		ass.getItemListeners()[0].itemStateChanged(null);
-
-		fc = new JCheckBox(Messages.getString("FonconfigEmbeddedFonts"), configuration.isMencoderFontConfig());
-		fc.setContentAreaFilled(false);
-		fc.addItemListener(new ItemListener() {
-			@Override
-			public void itemStateChanged(ItemEvent e) {
-				configuration.setMencoderFontConfig(e.getStateChange() == ItemEvent.SELECTED);
-			}
-		});
-		builder.add(GuiUtil.getPreferredSizeComponent(fc), FormLayoutUtil.flip(cc.xyw(3, 23, 5), colSpec, orientation));
-
-		builder.addLabel(Messages.getString("DvdvobsubSubtitlesQuality"), FormLayoutUtil.flip(cc.xy(1, 29), colSpec, orientation));
-		subq = new JTextField(configuration.getMencoderVobsubSubtitleQuality());
-		subq.addKeyListener(new KeyAdapter() {
-			@Override
-			public void keyReleased(KeyEvent e) {
-				configuration.setMencoderVobsubSubtitleQuality(subq.getText());
-			}
-		});
-		builder.add(subq, FormLayoutUtil.flip(cc.xyw(3, 29, 1), colSpec, orientation));
-
-		configuration.addConfigurationListener(new ConfigurationListener() {
-			@Override
-			public void configurationChanged(ConfigurationEvent event) {
-				if (event.getPropertyName() == null) {
-					return;
-				}
-				if ((!event.isBeforeUpdate()) && event.getPropertyName().equals(PmsConfiguration.KEY_DISABLE_SUBTITLES)) {
-					boolean enabled = !configuration.isDisableSubtitles();
-					ass.setEnabled(enabled);
-					fc.setEnabled(enabled);
-					mencoderNoassScale.setEnabled(enabled);
-					mencoderNoassOutline.setEnabled(enabled);
-					mencoderNoassBlur.setEnabled(enabled);
-					mencoderNoassSubpos.setEnabled(enabled);
-					ocw.setEnabled(enabled);
-					och.setEnabled(enabled);
-					subq.setEnabled(enabled);
-
-					if (enabled) {
-						ass.getItemListeners()[0].itemStateChanged(null);
-					}
-				}
-			}
-		});
-
-		JPanel panel = builder.getPanel();
-
-		// Apply the orientation to the panel and all components in it
-		panel.applyComponentOrientation(orientation);
-
-		return panel;
 	}
 
 	@Override
@@ -2514,7 +2116,7 @@ public class MEncoderVideo extends Player {
 		return Format.VIDEO;
 	}
 
-	private static String[] getSpecificCodecOptions(
+	public static String[] getSpecificCodecOptions(
 		String codecParam,
 		DLNAMediaInfo media,
 		OutputParams params,
