+++ src/main/java/net/pms/encoders/YoutubeDl.java
@@ -1,6 +1,5 @@
 /*
- * PS3 Media Server, for streaming any medias to your PS3.
- * Copyright (C) 2008-2012 A.Brochard
+ * This file is part of Universal Media Server, based on PS3 Media Server.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -24,7 +23,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import javax.swing.JComponent;
 import net.pms.configuration.DeviceConfiguration;
 import net.pms.configuration.PmsConfiguration;
 import net.pms.dlna.DLNAMediaInfo;
@@ -46,11 +44,6 @@ public class YoutubeDl extends FFMpegVideo {
 
 	// Not to be instantiated by anything but PlayerFactory
 	YoutubeDl() {
-	}
-
-	@Override
-	public JComponent config() {
-		return null;
 	}
 
 	@Override
