+++ src/main/java/net/pms/newgui/players/FFmpegAudio.java
@@ -1,0 +1,62 @@
+/*
+ * This file is part of Universal Media Server, based on PS3 Media Server.
+ *
+ * This program is a free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; version 2
+ * of the License only.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package net.pms.newgui.players;
+
+import com.jgoodies.forms.builder.PanelBuilder;
+import com.jgoodies.forms.factories.Borders;
+import com.jgoodies.forms.layout.CellConstraints;
+import com.jgoodies.forms.layout.FormLayout;
+import java.awt.Font;
+import java.awt.event.ItemEvent;
+import javax.swing.JCheckBox;
+import javax.swing.JComponent;
+import net.pms.Messages;
+import net.pms.PMS;
+import net.pms.configuration.PmsConfiguration;
+import net.pms.newgui.GuiUtil;
+
+public class FFmpegAudio {
+	private static final PmsConfiguration CONFIGURATION = PMS.getConfiguration();
+	private static JCheckBox noresample;
+
+	public static JComponent config() {
+		FormLayout layout = new FormLayout(
+			"left:pref, 0:grow",
+			"p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, p, 3dlu, 0:grow"
+		);
+		PanelBuilder builder = new PanelBuilder(layout);
+		builder.border(Borders.EMPTY);
+		builder.opaque(false);
+
+		CellConstraints cc = new CellConstraints();
+
+		JComponent cmp = builder.addSeparator(Messages.getString("GeneralSettings_SentenceCase"), cc.xyw(2, 1, 1));
+		cmp = (JComponent) cmp.getComponent(0);
+		cmp.setFont(cmp.getFont().deriveFont(Font.BOLD));
+
+		noresample = new JCheckBox(Messages.getString("AutomaticAudioResampling"), CONFIGURATION.isAudioResample());
+		noresample.setContentAreaFilled(false);
+		noresample.addItemListener((ItemEvent e) -> {
+			CONFIGURATION.setAudioResample(e.getStateChange() == ItemEvent.SELECTED);
+		});
+		builder.add(GuiUtil.getPreferredSizeComponent(noresample), cc.xy(2, 3));
+
+		return builder.getPanel();
+	}
+
+}
